lolcoach/requirements.txt
# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
sqlmodel>=0.0.22

# AI / RAG
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.1.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
chromadb>=0.6.0

# Security & Auth
keyring>=25.0.0
slowapi>=0.1.9

# Scraping & Ingestion
httpx>=0.28.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
PyMuPDF>=1.25.0

# Scheduling
apscheduler>=3.10.0

# System tray
pystray>=0.19.0
Pillow>=11.0.0

# Build
pyinstaller>=6.0.0
